+++ fake-cucumber/javascript/test/StepDefinitionRegistryTest.ts
@@ -3,8 +3,19 @@ import { messages } from 'cucumber-messages'
 import { stubMatchingStepDefinition } from './TestHelpers'
 
 import StepDefinitionRegistry from '../src/StepDefinitionRegistry'
+import gherkin from 'gherkin'
+import FakeTestResultsStream from '../src/FakeTestResultsStream'
+import { Readable } from 'stream'
 
 describe('StepDefinitionRegistry', () => {
+  describe('makeTestPlan', () => {
+    it('builds a plan from pickles', () => {
+      const stepDef = stubMatchingStepDefinition()
+
+      const registry = new StepDefinitionRegistry([stepDef])
+    })
+  })
+
   describe('#toMessages', () => {
     it('wraps each stepDefinitions.toMessages in an Envelope', () => {
       const stepDef1 = stubMatchingStepDefinition()
