+++ fake-cucumber/javascript/test/TestPlanTest.ts
@@ -1,0 +1,3 @@
+describe('TestPlan', () => {
+  // no-op
+})
